+++ src/components/sequence/items/WaitUntilAboveHorizonItem.vue
@@ -25,7 +25,7 @@
       <div class="text-xs text-slate-400 font-medium">
         {{ $t('components.sequence.items.center.ra') }}
       </div>
-      <div class="grid grid-cols-1 sm:grid-cols-3 gap-1">
+      <div class="grid grid-cols-1 @[20rem]:grid-cols-3 gap-1">
         <NumberInputPicker
           :modelValue="c.RAHours"
           label="h"
@@ -66,7 +66,7 @@
       <div class="text-xs text-slate-400 font-medium pt-1">
         {{ $t('components.sequence.items.center.dec') }}
       </div>
-      <div class="grid grid-cols-1 sm:grid-cols-3 gap-1">
+      <div class="grid grid-cols-1 @[20rem]:grid-cols-3 gap-1">
         <NumberInputPicker
           :modelValue="decDeg"
           label="°"